+++ src/main/java/ai/sahaj/feeModel/FeeModel.java
@@ -3,7 +3,7 @@ package ai.sahaj.feeModel;
 import ai.sahaj.VehicleType;
 
 public class FeeModel {
-    public double fees(VehicleType vehicleType, long hours) {
+    public int fees(VehicleType vehicleType, int hours) {
         return hours * vehicleType.rate;
     }
 }
